# perf: calculate grad on-the-fly for SiLUT